+++ inlang/source-code/editor/src/interface/editor/Footer.tsx
@@ -3,7 +3,45 @@ import IconGithub from "~icons/cib/github"
 import IconDiscord from "~icons/cib/discord"
 import Link from "#src/renderer/Link.jsx"
 import IconQuestionMark from "~icons/material-symbols/question-mark"
-// import IconExpandMore from "~icons/material-symbols/expand-more"
+
+export const getFinkResourcesLinks = () => {
+	return [
+		{
+			name: "About Fink",
+			href: import.meta.env.PROD
+				? "https://inlang.com/m/tdozzpar"
+				: "http://localhost:3000/m/tdozzpar",
+		},
+		{
+			name: "User Guide",
+			href: import.meta.env.PROD
+				? "https://inlang.com/g/6ddyhpoi"
+				: "http://localhost:3000/g/6ddyhpoi",
+		},
+		{
+			name: "About the ecosystem",
+			href: import.meta.env.PROD
+				? "https://inlang.com/documentation"
+				: "http://localhost:3000/documentation",
+		},
+		{
+			name: "Support Forum",
+			href: "https://discord.gg/gdMPPWy57R",
+		},
+		{
+			name: "Report a Bug",
+			href: "https://github.com/opral/inlang-fink/issues/new",
+		},
+		{
+			name: "Feature Request",
+			href: "https://github.com/opral/monorepo/discussions/categories/-fink-general",
+		},
+		{
+			name: "Submit Feedback",
+			href: "https://github.com/orgs/opral/discussions/categories/-fink-general",
+		}
+	]
+}
 
 const Footer = () => {
 	const socialMediaLinks = [
@@ -27,54 +65,11 @@ const Footer = () => {
 		},
 	]
 
-	const getResourcesLinks = () => {
-		return [
-			{
-				name: "About Fink",
-				href: import.meta.env.PROD ? "https://inlang.com/m/tdozzpar" : "http://localhost:3000/m/tdozzpar",
-			},
-			{
-				name: "User Guide",
-				href: import.meta.env.PROD ? "https://inlang.com/g/6ddyhpoi" : "http://localhost:3000/g/6ddyhpoi",
-			},
-			{
-				name: "About the ecosystem",
-				href: import.meta.env.PROD ? "https://inlang.com/documentation" : "http://localhost:3000/documentation",
-			},
-			{
-				name: "Support Forum",
-				href: "https://discord.gg/gdMPPWy57R",
-			},
-			{
-				name: "Report a Bug",
-				href: "https://github.com/opral/inlang-fink/issues/new",
-			},
-			{
-				name: "Feature Request",
-				href: "https://github.com/opral/monorepo/discussions/categories/-fink-general",
-			},
-			{
-				name: "Submit Feedback",
-				href: "https://github.com/orgs/opral/discussions/categories/-fink-general",
-			},
-			{
-				name: "Contact Us",
-				href: "mailto:hello@inlang.com",
-			},
-			{
-				name: "Join the Team",
-				href: "https://github.com/opral/monorepo/tree/main/careers",
-			}
-		]
-	}
-
 	return (
 		<footer class="bg-background border-t border-surface-200 py-4 px-4 overflow-visible">
 			<div class="max-w-7xl mx-auto flex justify-between gap-4 xl:px-4 h-6">
 				<div class="flex items-center justify-between">
-					<p class="text-sm text-surface-500">
-						© {new Date().getFullYear().toString()} Opral
-					</p>
+					<p class="text-sm text-surface-500">© {new Date().getFullYear().toString()} Opral</p>
 				</div>
 				<div class="flex gap-4 mr-[67.21px]">
 					<For each={socialMediaLinks}>
@@ -91,26 +86,22 @@ const Footer = () => {
 					</For>
 				</div>
 				<div class="relative">
-					<sl-dropdown
-						prop:placement="top-end"
-						prop:distance={4}
-						class="peer"
-					>
+					<sl-dropdown prop:placement="top-end" prop:distance={4} class="peer">
 						<button slot="trigger" class="bg-surface-900 rounded-full p-1">
 							<IconQuestionMark class="w-4 h-4 text-background" />
 						</button>
 						<sl-menu class="w-fit">
-							<For each={getResourcesLinks()}>
+							<For each={getFinkResourcesLinks()}>
 								{(link) => (
 									<>
 										<sl-menu-item>
-											<a
-												href={link.href} target="_blank"
-											>
+											<a href={link.href} target="_blank">
 												{link.name}
 											</a>
 										</sl-menu-item>
-										<Show when={link.name === "About the ecosystem" || link.name === "Submit Feedback"}>
+										<Show
+											when={link.name === "About the ecosystem"}
+										>
 											<div class="w-full border-b border-surface-200 my-1" />
 										</Show>
 									</>