+++ inlang/source-code/end-to-end-tests/paraglide-js/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @inlang/paraglide-js-e2e
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [2428451]
+  - @inlang/paraglide-js@1.5.0
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [d47b2aa]
+- Updated dependencies [192fdec]
+- Updated dependencies [0b7c82e]
+  - @inlang/paraglide-js@1.4.0
+
 ## 0.0.2
 
 ### Patch Changes