+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/pages/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @inlang/paraglide-js-adapter-next-example-pages
 
+## 0.2.14
+
+### Patch Changes
+
+- Updated dependencies [2428451]
+  - @inlang/paraglide-js@1.5.0
+  - @inlang/paraglide-js-adapter-next@3.0.0
+
+## 0.2.13
+
+### Patch Changes
+
+- Updated dependencies [d47b2aa]
+- Updated dependencies [192fdec]
+- Updated dependencies [0b7c82e]
+  - @inlang/paraglide-js@1.4.0
+  - @inlang/paraglide-js-adapter-next@2.0.0
+
 ## 0.2.12
 
 ### Patch Changes