+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/turbo/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @inlang/paraglide-js-adapter-next-example-turbo
 
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies [2428451]
+  - @inlang/paraglide-js@1.5.0
+
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies [d47b2aa]
+- Updated dependencies [192fdec]
+- Updated dependencies [0b7c82e]
+  - @inlang/paraglide-js@1.4.0
+
 ## 0.1.12
 
 ### Patch Changes